Use delay from pattern to postpone vibrate command
So far the `vibrate()` command was passed the complete `pattern` and used the
first parameter to delay the vibration. This led to the call itself colliding
with the one triggered by the system notification.

This commit extracts the `delay` parameter, uses it to postpone the call to
`vibrate()` itself and sets the first parameter of the `pattern` to zero to have
`vibrate()` trigger the vibration immediately when it's eventually called.